ng_ndp: prepare for parallel NDP implementations
