+++ sys/Makefile
@@ -49,6 +49,12 @@ endif
 ifneq (,$(filter ng_ndp,$(USEMODULE)))
     DIRS += net/network_layer/ng_ndp
 endif
+ifneq (,$(filter ng_ndp_internal,$(USEMODULE)))
+    DIRS += net/network_layer/ng_ndp/internal
+endif
+ifneq (,$(filter ng_ndp_node,$(USEMODULE)))
+    DIRS += net/network_layer/ng_ndp/node
+endif
 ifneq (,$(filter ng_netapi,$(USEMODULE)))
     DIRS += net/crosslayer/ng_netapi
 endif
