+++ sys/net/network_layer/ng_ndp/internal/Makefile
@@ -1,0 +1,3 @@
+MODULE = ng_ndp_internal
+
+include $(RIOTBASE)/Makefile.base
