+++ sys/net/network_layer/ng_ndp/node/Makefile
@@ -1,0 +1,3 @@
+MODULE = ng_ndp_node
+
+include $(RIOTBASE)/Makefile.base
